Don't send memory_slots on the wire
The memory_slots parameter isn't part of the expected aqt payload. While
it doesn't cause a failure it's not good to include extra fields in the
json payload, they might start rejecting unknown fields in the future.
The memory_slots are strictly needed for the qiskit result object to
ensure we're padding our response correctly. This commit updates the
qobj_to_aqt function to not include the memory slots and instead updates
the backend to include the qobj in the AQTJob object creation. This
gives the AQTJob class full access to all the parameters in the qobj,
which both enables us to set the memory_slots and also the qobj_id
fields in the output Result objects.