Fix tests for qobj_to_aqt() function
In the recent changes to how the qobj_to_aqt() function works the
unittests verifying it's functionality hasn't kept up to date. This is
mostly because we don't have any CI setup yet to check commits as they
come in. This commit fixes the issues with the tests to make sure that
they work and verify the current functionality of the function.