Marginally improve performance of bit rearrangement
This commit makes some small improvements for performance to the bit
rearrangement step for generating a Result object from a job. The main
thing is it moves the qubit mapping from being generated for each shot
(which can be 250 times) to just happening once when the job is created.
The mapping is static based on the instructions in the qobj (which qubit
is measured to which clbit) so we don't need to dynamically figure it
out for each result in the list. The second thing is mostly trivial, we
don't need to call reverse() on a list of all 0s when we generate that
list as the first step in rearranging. They're all identical values it
will be the same reversed or not.