Set name in result header
Running quantum volume from ignis requires that the experiment name from
the input qobj/circuit gets set in the result header (it's used for
grouping the results data) however the AQTJob class was not setting
this in the generated result object. This commit adds the name to the
result header from the input qobj experiment to match ignis's
expectations.