Change version number to be a beta release
We want to push out an early release to get the package name on pypi
registered and test the release machinary. However, because of the
dependency on unreleased terra features it's probably not the best idea
to mark it as a full release. This changes the version number to be
0.0.1b1 to indicate it's the first beta of the first release. This
should let us register the package but indicate it's not ready yet.