Add dynamic version info to version string
This commit adds dynamic version info to version string for the package.
Previously the version string was just hardcoded in the __init__.py for
the package and the setup.py. Instead this moves to have a single
version text file that contains the current version to be used for
packaging and the __version__ attribute. But it also will update the
__version__ attribute to properly report if you're running a release or
a dev version (including a short sha1).