Small cleanups to the project
This commit adjust some small minor details to the project. The first
and largest is that it bumps the minimum terra version. This is because
0.10.0 has been released (for some time now) which includes the initial
support for transpiling to the ion trap basis gate set. So we should set
that version as the minimum. The associated readme line is updated to
reflect that. The rest of the changes are mostly just docstring,
whitespace, and comment fixes to tidy up the code slightly.