Bump pin to unify lint job versions across elements (#12)
In an effort to unify the versions of pylint run across all the elements
this commit bumps the pinned versions of pylint and astroid to be the
latest and what we are syncing all the elements to use. This should make
working between all the elements easier to do. We should have been
pinning the pylint version anyway because each release of pylint and
astroid tends to introduce new failures.

Related To: Qiskit/qiskit-terra#3629