Update the package metadata in setup.py
This commit updates the package metadata prior to release. First it adds
the README.md file as the long_description to the package metadata. This
means that on pypi the README will show up on the package page. Then it
adds the project_urls field which is used to provide specific links to
the code repo and the issue tracker for the project.