Update package metadata and README for new location (#19)
The qiskit github organization is in the process of being reorganized to
concentrate on the core elements, dependencies, development tools, etc. As
part of this reoganization 3rd party providers are being moved to projects
under the qiskit-community github organization. As part of this the
qiskit-aqt-provider repository was moved to qiskit-community github
organization. This commit updates the README and package metadata to
reflect this change.