Drop Python 3.5 support (#23)
Python 3.5 went EoL in September of this year and the from the
Qiskit Terra 0.16.0 release onwards Python 3.5 is no longer supported in
terra. Since future work on the aqt provider will depend on newer
versions of Qiskit Terra we need to be able to use the later versions of
Terra. This commit drops support for python 3.5 from the aqt provider,
from this point onwards python 3.6 will be the minimum version of python
that works.