Use new provider interface (#22)
* Use new provider interface

In qiskit-terra 0.16.0 a new version of the provider interface was
added. This version is now explicitly versioned to enable graceful
changes to the interface over time. For the first version of this new
interface the key changes are that circuits can be passed directly to
Backend.run() and to go along with that a new options object has been
added to the backend to track run time options for a backend. This
commit updates the internal aqt provider interface to use the new
version instead of the previous legacy providers interface.

* Fix stacklevel for option warning and remove stray print