Emit a single warning about pending package namespace change
In the release after next, 0.4.0, we will change the namespace of the
package from 'qiskit.providers.aqt' to 'qiskit_aqt_provider'. This is
being done for 2 reasons. The first is to make it clear what is developed
as part of the core qiskit project, and things on top of it. The second
is that python namespace packaging is quite error prone and a frequent
cause of user issues. So having the aqt provider package own its own
namespace should make it much more reliable to install and use. (see
Qiskit/qiskit#559 for more details). While the first reason may change
over time as the priorities and objective of the Qiskit project evolve,
the second will always remain true. So even if the first goes away this
will be a one way permanent change.