Add unittest for full results processing
This commit adds a unittest to validate the result processing builds the
expected counts dictionary from a full path execution. It uses a
captured response json from the AQT api to mock out the job method to
get the result data and uses that to validate the provider side
processing builds a result object as expected.