Fix docs artifact upload path (#41)
In #40 the ci configuration was changed to use github actions, part of
what that offers is an artifact store where we can upload the built docs
for download to check the output. However there was a typo in that job
configuration and it trying to upload the incorrect path. This commit
corrects that oversight so that we will upload the built docs at the end
of every docs job.