Preserve circuit metadata in output result (#48)
This commit fixes an issue with the Result objects generated by the
AqtJob. Previously the metadata from the input circuit was not passed to
the result header. This causes applications (like Qiskit Experiments) to
not be able to use the circuit metadata for tracking purposes. This
commit fixes this oversight by adding any circuit metadata to the output
result header.