Migrate AQT backends to BackendV2 (#51)
* Migrate AQT backends to BackendV2

In qiskit-terra 0.19.0 a new version of the backend interface,
BackendV2, was introduced. This interface provides a new simpler
interface for describing backends to users and the qiskit compiler. This
commit migrates the AQT provider's backend objects to use this new
interface. As part of this migration old deprecated functionality from
the BackendV1 migration last year has been removed and functionality no
longer exposed as part of BackendV2 is deprecated.

* Bump minimum terra version

* Fix lint errors

* Deprecate empty properties() method

* Fix doc syntax issue