Remove outdated docs template and CSS overrides (#55)
These HTML template overrides mean that you won't be getting improvements to qiskit_sphinx_theme. They also risk breaks because we assume the files are not overridden. It can be valid to override the templates, but should only be done intentionally and with an understanding that increases your maintenance burden.

The custom.css file was not ever activated, and also only has rules for .toggle which is no longer used.

This also:

    Adds missing pyproject.toml, which is strongly encouraged to have for Python packaging.
    Activates jQuery, which is no longer activated by default.
    Pins the qiskit_sphinx_theme by minor version. While we try to keep breaking changes to a minimum, we are undergoing big improvements to the theme right now so this gives you better stability.

* Remove outdated docs template overrides

* Also remove unused custom.css

* Loosen Sphinx requirement

* Cherry pick https://github.com/Qiskit-Partners/qiskit-aqt-provider/pull/56